initial pass, switch crosswalk
